+++ serde-version/src/deserializer.rs
@@ -1,29 +1,34 @@
 use super::visitor::VersionedVisitor;
 use super::Error;
+use failure::_core::borrow::Borrow;
 use serde::Deserializer;
 use std::collections::HashMap;
+use std::hash::Hash;
 
 /// Maps the version number for each deserialization type name
-pub type VersionMap = HashMap<String, usize>;
+pub trait VersionMap {
+    fn get(&self, type_id: &str) -> Option<usize>;
+}
+pub type DefaultVersionMap = HashMap<String, usize>;
 
 /// A wrapper around a deserialize to support the deserialization.
 ///
 /// This deserializer will wrap all calls where specialization is required. (Like
 /// `next_element`, `next_value`, ...)
-pub struct VersionedDeserializer<'de, D>
+pub struct VersionedDeserializer<'de, D, VM>
 where
     D: Deserializer<'de>,
 {
     deserializer: D,
-    version_map: &'de VersionMap,
+    version_map: &'de VM,
     marker: std::marker::PhantomData<&'de usize>,
 }
 
-impl<'de, D> VersionedDeserializer<'de, D>
+impl<'de, D, VM> VersionedDeserializer<'de, D, VM>
 where
     D: Deserializer<'de>,
 {
-    pub fn new(deserializer: D, version_map: &'de VersionMap) -> Self {
+    pub fn new(deserializer: D, version_map: &'de VM) -> Self {
         Self {
             deserializer,
             version_map,
@@ -47,7 +52,9 @@ macro_rules! forward_deserialize {
     }
 }
 
-impl<'de, D: Deserializer<'de>> Deserializer<'de> for VersionedDeserializer<'de, D> {
+impl<'de, D: Deserializer<'de>, VM: VersionMap> Deserializer<'de>
+    for VersionedDeserializer<'de, D, VM>
+{
     type Error = Error<D::Error>;
 
     forward_deserialize!(deserialize_any);
@@ -83,4 +90,22 @@ impl<'de, D: Deserializer<'de>> Deserializer<'de> for VersionedDeserializer<'de,
                          name => &'static str,
                          variants => &'static [&'static str]);
     forward_deserialize!(deserialize_ignored_any);
+}
+
+impl<T: Borrow<str> + Hash + Eq> VersionMap for HashMap<T, usize> {
+    fn get(&self, type_id: &str) -> Option<usize> {
+        std::collections::HashMap::get(self, type_id).cloned()
+    }
+}
+
+impl<'a, T: Borrow<str> + Hash + Eq> VersionMap for &'a HashMap<T, usize> {
+    fn get(&self, type_id: &str) -> Option<usize> {
+        std::collections::HashMap::get(self, type_id).cloned()
+    }
+}
+
+impl<'a, T: Borrow<str> + Hash + Eq> VersionMap for &'a mut HashMap<T, usize> {
+    fn get(&self, type_id: &str) -> Option<usize> {
+        std::collections::HashMap::get(self, type_id).cloned()
+    }
 }